requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
pkg_resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
